Psybot/Backend/new-env/requirements.txt
numpy
Pillow
flask
flask_cors
requests
python-dotenv
